Handle errors from url-queue-retrieve
Previously, if the url-queue job encountered an error, erc-image's
callback function ignored the error.  Because the job buffer was empty,
the search-forward function call in the callback function errored out.
The error from the callback function prevented url-queue from removing
the job from the progress queue.  As a result, url-queue kept endlessly
calling the callback function for the failed job.

* erc-image.el (erc-image-insert-other-buffer, erc-image-insert-inline):
Check the status, and if there was an error, print it and return.