Fix rescaling
Now allows the internal image manipulation introduced in 27.1. ImageMagick is now only needed before versions prior to 27.1. This allows rescaling to work in MS Windows.

Now setting erc-image-inline-rescale to a number only scales down. Previously, setting erc-image-inline-rescale to a number always resized an image to that height even if it was smaller. The width was ignored entirely. For example, if we `(setq erc-image-inline-rescale 400)`, then a 100x50 image would be scaled up to 800x400. This behavior is inconsistent with `(setq erc-image-inline-rescale 'window)` which only scaled down if the image exceeded window height or width. Additionally, it has the undesirable side effect of possibly making an image wider than the window.

New internal function to eliminate repetition. Since the same logic is now used for both the  `(numberp erc-image-inline-rescale)` and `(eq erc-image-inline-rescale 'window)`, it's been split out into its own function `erc-image--maybe-rescale'